Refactor disease components so that RemoteDataTable is more reusable
